fixup TokenUpdateOptions > TokenAbstractOptions
