+++ tests/Backend/Snowflake/TimeTravelTest.php
@@ -5,6 +5,7 @@ namespace Keboola\Test\Backend\Snowflake;
 use Keboola\Csv\CsvFile;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
+use Keboola\StorageApi\Options\TokenAbstractOptions;
 use Keboola\StorageApi\Options\TokenCreateOptions;
 use Keboola\StorageApi\Options\TokenUpdateOptions;
 use Keboola\StorageApi\TableExporter;
@@ -199,7 +200,7 @@ class TimeTravelTest extends StorageApiTestCase
         // Setup our test clients
         $outputBucketTokenOptions = (new TokenCreateOptions())
             ->setDescription('Output bucket only write token')
-            ->addBucketPermission($this->getTestBucketId(self::STAGE_OUT), TokenUpdateOptions::BUCKET_PERMISSION_WRITE)
+            ->addBucketPermission($this->getTestBucketId(self::STAGE_OUT), TokenAbstractOptions::BUCKET_PERMISSION_WRITE)
         ;
 
         $outputBucketTokenId = $this->_client->createToken($outputBucketTokenOptions);
@@ -216,7 +217,7 @@ class TimeTravelTest extends StorageApiTestCase
 
         $inputBucketTokenOptions = (new TokenCreateOptions())
             ->setDescription('Input bucket only read token')
-            ->addBucketPermission($this->getTestBucketId(), TokenUpdateOptions::BUCKET_PERMISSION_READ)
+            ->addBucketPermission($this->getTestBucketId(), TokenAbstractOptions::BUCKET_PERMISSION_READ)
         ;
 
         $inputBucketTokenId = $this->_client->createToken($inputBucketTokenOptions);
@@ -233,8 +234,8 @@ class TimeTravelTest extends StorageApiTestCase
 
         $minimalTokenOptions = (new TokenCreateOptions())
             ->setDescription('Minimal permissions token')
-            ->addBucketPermission($this->getTestBucketId(), TokenUpdateOptions::BUCKET_PERMISSION_READ)
-            ->addBucketPermission($this->getTestBucketId(self::STAGE_OUT), TokenUpdateOptions::BUCKET_PERMISSION_WRITE)
+            ->addBucketPermission($this->getTestBucketId(), TokenAbstractOptions::BUCKET_PERMISSION_READ)
+            ->addBucketPermission($this->getTestBucketId(self::STAGE_OUT), TokenAbstractOptions::BUCKET_PERMISSION_WRITE)
         ;
 
         $minimalTokenId = $this->_client->createToken($minimalTokenOptions);
